Move Listenable implementation to serial module.
